tests: Moving all unit tests to the ./test directory
